move JobList files in one folder - extract two dialog components
