+++ services/orchest-webserver/client/src/jobs-view/JobsView.tsx
@@ -1,10 +1,10 @@
-import JobList from "@/components/JobList";
 import { Layout } from "@/components/Layout";
 import ProjectBasedView from "@/components/ProjectBasedView";
 import { useCustomRoute } from "@/hooks/useCustomRoute";
 import { useSendAnalyticEvent } from "@/hooks/useSendAnalyticEvent";
 import { siteMap } from "@/Routes";
 import React from "react";
+import JobList from "./JobList";
 
 const JobsView: React.FC = () => {
   const { projectUuid } = useCustomRoute();
